fix: no more underflows when combining spans

+++ src/lib.rs
@@ -25,14 +25,15 @@ trait Combine {
 impl Combine for SourceSpan {
 	fn combine(&self, other: &Self) -> Self {
 		// Start at the first span
-		let start = self.offset();
+		let start = self.offset() as isize;
 
 		// Keep going for the length of the first span, compensate for a
 		// potential gap between the first and second span, and keep going for
 		// the length of the second span
-		let first_len = self.len();
-		let len = first_len + (other.offset() - (start + first_len - 1)) + other.len();
+		let first_len = self.len() as isize;
+		let len =
+			first_len + (other.offset() as isize - (start + first_len - 1)) + other.len() as isize;
 
-		(start, len).into()
+		(start as usize, len as usize).into()
 	}
 }
